requirements.txt
# 期权套利扫描工具 - 依赖包清单
# Option Arbitrage Scanner - Dependencies

# 核心框架
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# 数据获取
tushare>=1.2.89
requests>=2.31.0
aiohttp>=3.8.0

# 数学计算
scipy>=1.10.0
scikit-learn>=1.3.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 缓存
redis>=4.5.0
diskcache>=5.6.0

# 日志
loguru>=0.7.0

# 图表展示和可视化
plotly>=5.15.0
altair>=5.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 数据导出和报告生成
openpyxl>=3.1.0
reportlab>=4.0.0
xlsxwriter>=3.1.0

# UI增强组件
streamlit-tags>=1.2.8

# 图像处理（用于图表导出）
pillow>=10.0.0

# 工具库
typing-extensions>=4.7.0
pydantic>=2.0.0
click>=8.1.0

# 开发工具
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0